Enum cynic_codegen::RenameAll

pub enum RenameAll {
    None,
    Lowercase,
    Uppercase,
    PascalCase,
    CamelCase,
    SnakeCase,
    ScreamingSnakeCase,
}

Rules to rename all fields in an InputObject or variants in an Enum as GraphQL naming conventions usually don’t match rust

Variants

None

Lowercase

For names that are entirely lowercase in GraphQL: myfield

Uppercase

For names that are entirely uppercase in GraphQL: MYFIELD

PascalCase

For names that are entirely pascal case in GraphQL: MyField

CamelCase

For names that are entirely camel case in GraphQL: myField

SnakeCase

For names that are entirely snake case in GraphQL: my_field

ScreamingSnakeCase

For names that are entirely snake case in GraphQL: MY_FIELD